+++ src/js/functions.js
@@ -63,3 +63,12 @@ function clamp(value, min, max) {
 function isDigit(char) {
     return /^[0-9]$/.test(char);
 }
+
+function getFormattedDate() {
+    const now = new Date();
+    const year = String(now.getFullYear()).slice(2);
+    const month = String(now.getMonth() + 1).padStart(2, "0");
+    const day = String(now.getDate()).padStart(2, "0");
+
+    return `${year}-${month}-${day}`;
+}